+++ Manifiestapp/src/app/pages/speakers/speaker-detail.page.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { ISpeaker } from 'src/app/shared/models/Event.interface';
+import { forkJoin } from 'rxjs';
+import { IEvent, ISpeaker } from 'src/app/shared/models/Event.interface';
 import { ProgrammeService } from 'src/app/shared/services/data/programme/programme.service';
 
 @Component({
@@ -12,6 +13,7 @@ export class SpeakerDetailPage {
     id: string;
     speaker: ISpeaker;
     defaultHref = '/speakers';
+    linkedEvents: IEvent[] = [];
 
     constructor(
         private programmeService: ProgrammeService,
@@ -22,6 +24,11 @@ export class SpeakerDetailPage {
         this.id = this.activatedRoute.snapshot.params.id;
         this.programmeService.getSpeaker(this.id).subscribe(data => {
             this.speaker = data;
+            forkJoin(this.speaker.related_events.map(sre => this.programmeService.getEventsByParentId(sre.uuid))).subscribe(events => {
+                try {
+                    events.forEach(e => e.forEach(ee => this.linkedEvents.push(ee)))
+                } catch (error) { }
+            });
         });
     }